+++ blocks/vue-blocks/src/pqeaVueApp.vue
@@ -99,7 +99,7 @@
 
             <!-- Results Loop -->
             <template v-if="pqeas.length > 0" v-for="(pqea, index) in paginatedPqeas" :key="index">
-                <tr :class="`pqea result result--${index+1} ${0 === (index+1) % 2 ? `even` : `odd`}`" tabindex="0">
+                <tr :class="`pqea result result--${index+1} ${0 === (index+1) % 2 ? `even` : `odd`}`">
                     <!-- Company Name and Head Office -->
                     <td data-label="Company Name and Head Office" class="pqea__company-and-location">
                         <!-- Company Website Link -->